+++ roborock/devices/traits/b01/__init__.py
@@ -6,8 +6,7 @@ from roborock.devices.mqtt_channel import MqttChannel
 from roborock.devices.traits import Trait
 from roborock.roborock_message import RoborockB01Props
 
-__init__ = [
-    "create_b01_traits",
+__all__ = [
     "PropertiesApi",
 ]