+++ roborock/devices/traits/v1/valley_electricity_timer.py
@@ -18,8 +18,7 @@ class ValleyElectricityTimerTrait(ValleyElectricityTimer, common.V1TraitMixin, c
 
     async def set_timer(self, timer: ValleyElectricityTimer) -> None:
         """Set the Valley Electricity Timer settings of the device."""
-        await self.rpc_channel.send_command(RoborockCommand.SET_VALLEY_ELECTRICITY_TIMER, params=timer.as_dict())
-        await self.refresh()
+        await self.rpc_channel.send_command(RoborockCommand.SET_VALLEY_ELECTRICITY_TIMER, params=timer.as_list())
 
     async def clear_timer(self) -> None:
         """Clear the Valley Electricity Timer settings of the device."""
@@ -30,20 +29,13 @@ class ValleyElectricityTimerTrait(ValleyElectricityTimer, common.V1TraitMixin, c
         """Enable the Valley Electricity Timer settings of the device."""
         await self.rpc_channel.send_command(
             RoborockCommand.SET_VALLEY_ELECTRICITY_TIMER,
-            params={
-                **self.as_dict(),
-                _ENABLED_PARAM: 1,
-            },
+            params=self.as_list(),
         )
         self.enabled = 1
 
     async def disable(self) -> None:
         """Disable the Valley Electricity Timer settings of the device."""
         await self.rpc_channel.send_command(
-            RoborockCommand.SET_VALLEY_ELECTRICITY_TIMER,
-            params={
-                **self.as_dict(),
-                _ENABLED_PARAM: 0,
-            },
+            RoborockCommand.CLOSE_VALLEY_ELECTRICITY_TIMER,
         )
         self.enabled = 0